requirements.txt
openai
faiss-cpu
scikit-learn
lightgbm
xgboost
python-dotenv
pandas
numpy
tqdm
sentence-transformers
